Hive/requirements.txt
# Core dependencies
mcp>=0.1.0
fastapi>=0.95.0
uvicorn>=0.21.1
websockets>=11.0.1
python-dotenv>=1.0.0
anyio>=3.6.2

# LLM providers
google-generativeai>=0.4.0
groq>=0.4.0

# Database
sqlalchemy>=2.0.0

# HTTP client
requests>=2.28.2
httpx>=0.24.0
sseclient>=0.0.27

# Utilities
pydantic>=2.0.0 